Add Poseidon fuzzer, fix conversion warning
- New fuzz target for Poseidon hash (1-8 inputs)
- Fix implicit int conversion warning in threshold swap

+++ native/tetsuo-core/src/agenc_zk.c
@@ -142,7 +142,7 @@ agenc_zk_result_t agenc_zk_verify(
     }
 
     /* Check threshold matches */
-    uint16_t proof_threshold = (proof->threshold >> 8) | (proof->threshold << 8); /* big-endian */
+    uint16_t proof_threshold = (uint16_t)((proof->threshold >> 8) | (proof->threshold << 8));
     if (proof_threshold < threshold) {
         return AGENC_ZK_ERR_BELOW_THRESHOLD;
     }
